+++ test/test_gromov.py
@@ -570,18 +570,106 @@ def test_entropic_gromov_dtype_device(nx):
 
         C1b, C2b, pb, qb = nx.from_numpy(C1, C2, p, q, type_as=tp)
 
-        for solver in ['PGD', 'PPA']:
-            Gb = ot.gromov.entropic_gromov_wasserstein(
-                C1b, C2b, pb, qb, 'square_loss', epsilon=1e-1, max_iter=5,
-                solver=solver, verbose=True
-            )
-            gw_valb = ot.gromov.entropic_gromov_wasserstein2(
-                C1b, C2b, pb, qb, 'square_loss', epsilon=1e-1, max_iter=5,
-                solver=solver, verbose=True
-            )
+        for solver in ['PGD', 'PPA', 'BAPG']:
+            if solver == 'BAPG':
+                Gb = ot.gromov.BAPG_gromov_wasserstein(
+                    C1b, C2b, pb, qb, max_iter=2, verbose=True)
+                gw_valb = ot.gromov.BAPG_gromov_wasserstein2(
+                    C1b, C2b, pb, qb, max_iter=2, verbose=True)
+            else:
+                Gb = ot.gromov.entropic_gromov_wasserstein(
+                    C1b, C2b, pb, qb, max_iter=2, solver=solver, verbose=True)
+                gw_valb = ot.gromov.entropic_gromov_wasserstein2(
+                    C1b, C2b, pb, qb, max_iter=2, solver=solver, verbose=True)
 
             nx.assert_same_dtype_device(C1b, Gb)
             nx.assert_same_dtype_device(C1b, gw_valb)
+
+
+def test_BAPG_gromov(nx):
+    n_samples = 10  # nb samples
+
+    mu_s = np.array([0, 0])
+    cov_s = np.array([[1, 0], [0, 1]])
+
+    xs = ot.datasets.make_2D_samples_gauss(n_samples, mu_s, cov_s, random_state=42)
+
+    xt = xs[::-1].copy()
+
+    p = ot.unif(n_samples)
+    q = ot.unif(n_samples)
+    G0 = p[:, None] * q[None, :]
+    C1 = ot.dist(xs, xs)
+    C2 = ot.dist(xt, xt)
+
+    C1 /= C1.max()
+    C2 /= C2.max()
+
+    C1b, C2b, pb, qb, G0b = nx.from_numpy(C1, C2, p, q, G0)
+
+    # complete test with marginal loss = True
+    marginal_loss = True
+    with pytest.raises(ValueError):
+        loss_fun = 'weird_loss_fun'
+        G, log = ot.gromov.BAPG_gromov_wasserstein(
+            C1, C2, None, q, loss_fun, symmetric=None, G0=G0,
+            epsilon=1e-1, max_iter=10, marginal_loss=marginal_loss,
+            verbose=True, log=True)
+
+    G, log = ot.gromov.BAPG_gromov_wasserstein(
+        C1, C2, None, q, 'square_loss', symmetric=None, G0=G0,
+        epsilon=1e-1, max_iter=10, marginal_loss=marginal_loss,
+        verbose=True, log=True)
+    Gb = nx.to_numpy(ot.gromov.BAPG_gromov_wasserstein(
+        C1b, C2b, pb, None, 'square_loss', symmetric=True, G0=None,
+        epsilon=1e-1, max_iter=10, marginal_loss=marginal_loss, verbose=True,
+        log=False
+    ))
+
+    # check constraints
+    np.testing.assert_allclose(G, Gb, atol=1e-06)
+    np.testing.assert_allclose(
+        p, Gb.sum(1), atol=1e-02)  # cf convergence gromov
+    np.testing.assert_allclose(
+        q, Gb.sum(0), atol=1e-02)  # cf convergence gromov
+
+    with pytest.warns(UserWarning):
+
+        gw = ot.gromov.BAPG_gromov_wasserstein2(
+            C1, C2, p, q, 'kl_loss', symmetric=False, G0=None,
+            max_iter=10, epsilon=1e-2, marginal_loss=marginal_loss, log=False)
+
+    gw, log = ot.gromov.BAPG_gromov_wasserstein2(
+        C1, C2, p, q, 'kl_loss', symmetric=False, G0=None,
+        max_iter=10, epsilon=1., marginal_loss=marginal_loss, log=True)
+    gwb, logb = ot.gromov.BAPG_gromov_wasserstein2(
+        C1b, C2b, pb, qb, 'kl_loss', symmetric=None, G0=G0b,
+        max_iter=10, epsilon=1., marginal_loss=marginal_loss, log=True)
+    gwb = nx.to_numpy(gwb)
+
+    G = log['T']
+    Gb = nx.to_numpy(logb['T'])
+
+    np.testing.assert_allclose(gw, gwb, atol=1e-06)
+    np.testing.assert_allclose(gw, 0, atol=1e-1, rtol=1e-1)
+
+    # check constraints
+    np.testing.assert_allclose(G, Gb, atol=1e-06)
+    np.testing.assert_allclose(
+        p, Gb.sum(1), atol=1e-02)  # cf convergence gromov
+    np.testing.assert_allclose(
+        q, Gb.sum(0), atol=1e-02)  # cf convergence gromov
+
+    marginal_loss = False
+    G, log = ot.gromov.BAPG_gromov_wasserstein(
+        C1, C2, None, q, 'square_loss', symmetric=None, G0=G0,
+        epsilon=1e-1, max_iter=10, marginal_loss=marginal_loss,
+        verbose=True, log=True)
+    Gb = nx.to_numpy(ot.gromov.BAPG_gromov_wasserstein(
+        C1b, C2b, pb, None, 'square_loss', symmetric=False, G0=None,
+        epsilon=1e-1, max_iter=10, marginal_loss=marginal_loss, verbose=True,
+        log=False
+    ))
 
 
 @pytest.skip_backend("tf", reason="test very slow with tf backend")
@@ -722,6 +810,99 @@ def test_entropic_proximal_fgw(nx):
         q, Gb.sum(0), atol=1e-04)  # cf convergence gromov
 
 
+def test_BAPG_fgw(nx):
+    n_samples = 5  # nb samples
+
+    mu_s = np.array([0, 0])
+    cov_s = np.array([[1, 0], [0, 1]])
+
+    xs = ot.datasets.make_2D_samples_gauss(n_samples, mu_s, cov_s, random_state=42)
+
+    xt = xs[::-1].copy()
+
+    rng = np.random.RandomState(42)
+    ys = rng.randn(xs.shape[0], 2)
+    yt = ys[::-1].copy()
+
+    p = ot.unif(n_samples)
+    q = ot.unif(n_samples)
+    G0 = p[:, None] * q[None, :]
+
+    C1 = ot.dist(xs, xs)
+    C2 = ot.dist(xt, xt)
+
+    C1 /= C1.max()
+    C2 /= C2.max()
+
+    M = ot.dist(ys, yt)
+
+    Mb, C1b, C2b, pb, qb, G0b = nx.from_numpy(M, C1, C2, p, q, G0)
+
+    with pytest.raises(ValueError):
+        loss_fun = 'weird_loss_fun'
+        G, log = ot.gromov.BAPG_fused_gromov_wasserstein(
+            M, C1, C2, p, q, loss_fun=loss_fun, max_iter=1, log=True)
+
+    # complete test with marginal loss = True
+    marginal_loss = True
+
+    G, log = ot.gromov.BAPG_fused_gromov_wasserstein(
+        M, C1, C2, p, q, 'square_loss', symmetric=None, G0=G0,
+        epsilon=1e-1, max_iter=10, marginal_loss=marginal_loss, log=True)
+    Gb = nx.to_numpy(ot.gromov.BAPG_fused_gromov_wasserstein(
+        Mb, C1b, C2b, pb, qb, 'square_loss', symmetric=True, G0=None,
+        epsilon=1e-1, max_iter=10, marginal_loss=marginal_loss, verbose=True))
+
+    # check constraints
+    np.testing.assert_allclose(G, Gb, atol=1e-06)
+    np.testing.assert_allclose(
+        p, Gb.sum(1), atol=1e-02)  # cf convergence gromov
+    np.testing.assert_allclose(
+        q, Gb.sum(0), atol=1e-02)  # cf convergence gromov
+
+    with pytest.warns(UserWarning):
+
+        fgw = ot.gromov.BAPG_fused_gromov_wasserstein2(
+            M, C1, C2, p, q, 'kl_loss', symmetric=False, G0=None,
+            max_iter=10, epsilon=1e-3, marginal_loss=marginal_loss, log=False)
+
+    fgw, log = ot.gromov.BAPG_fused_gromov_wasserstein2(
+        M, C1, C2, p, None, 'kl_loss', symmetric=True, G0=None,
+        max_iter=5, epsilon=1, marginal_loss=marginal_loss, log=True)
+    fgwb, logb = ot.gromov.BAPG_fused_gromov_wasserstein2(
+        Mb, C1b, C2b, None, qb, 'kl_loss', symmetric=None, G0=G0b,
+        max_iter=5, epsilon=1, marginal_loss=marginal_loss, log=True)
+    fgwb = nx.to_numpy(fgwb)
+
+    G = log['T']
+    Gb = nx.to_numpy(logb['T'])
+
+    np.testing.assert_allclose(fgw, fgwb, atol=1e-06)
+    np.testing.assert_allclose(fgw, 0, atol=1e-1, rtol=1e-1)
+
+    # check constraints
+    np.testing.assert_allclose(G, Gb, atol=1e-06)
+    np.testing.assert_allclose(
+        p, Gb.sum(1), atol=1e-02)  # cf convergence gromov
+    np.testing.assert_allclose(
+        q, Gb.sum(0), atol=1e-02)  # cf convergence gromov
+
+    # Tests with marginal_loss = False
+    marginal_loss = False
+    G, log = ot.gromov.BAPG_fused_gromov_wasserstein(
+        M, C1, C2, p, q, 'square_loss', symmetric=False, G0=G0,
+        epsilon=1e-1, max_iter=10, marginal_loss=marginal_loss, log=True)
+    Gb = nx.to_numpy(ot.gromov.BAPG_fused_gromov_wasserstein(
+        Mb, C1b, C2b, pb, qb, 'square_loss', symmetric=None, G0=None,
+        epsilon=1e-1, max_iter=10, marginal_loss=marginal_loss, verbose=True))
+    # check constraints
+    np.testing.assert_allclose(G, Gb, atol=1e-06)
+    np.testing.assert_allclose(
+        p, Gb.sum(1), atol=1e-02)  # cf convergence gromov
+    np.testing.assert_allclose(
+        q, Gb.sum(0), atol=1e-02)  # cf convergence gromov
+
+
 def test_asymmetric_entropic_fgw(nx):
     n_samples = 5  # nb samples
     rng = np.random.RandomState(0)
@@ -797,15 +978,18 @@ def test_entropic_fgw_dtype_device(nx):
 
         Mb, C1b, C2b, pb, qb = nx.from_numpy(M, C1, C2, p, q, type_as=tp)
 
-        for solver in ['PGD', 'PPA']:
-            Gb = ot.gromov.entropic_fused_gromov_wasserstein(
-                Mb, C1b, C2b, pb, qb, 'square_loss', epsilon=0.1, max_iter=5,
-                solver=solver, verbose=True
-            )
-            fgw_valb = ot.gromov.entropic_fused_gromov_wasserstein2(
-                Mb, C1b, C2b, pb, qb, 'square_loss', epsilon=0.1, max_iter=5,
-                solver=solver, verbose=True
-            )
+        for solver in ['PGD', 'PPA', 'BAPG']:
+            if solver == 'BAPG':
+                Gb = ot.gromov.BAPG_fused_gromov_wasserstein(
+                    Mb, C1b, C2b, pb, qb, max_iter=2)
+                fgw_valb = ot.gromov.BAPG_fused_gromov_wasserstein2(
+                    Mb, C1b, C2b, pb, qb, max_iter=2)
+
+            else:
+                Gb = ot.gromov.entropic_fused_gromov_wasserstein(
+                    Mb, C1b, C2b, pb, qb, max_iter=2, solver=solver)
+                fgw_valb = ot.gromov.entropic_fused_gromov_wasserstein2(
+                    Mb, C1b, C2b, pb, qb, max_iter=2, solver=solver)
 
             nx.assert_same_dtype_device(C1b, Gb)
             nx.assert_same_dtype_device(C1b, fgw_valb)